Feature/inventory service - Added stock check functionality and exception handling of insufficient stock

+++ order-service/src/main/java/com/orderprocessingeda/orderservice/entity/OrderItem.java
@@ -21,7 +21,7 @@ public class OrderItem {
     private Long productId;
 
     @Column(nullable = false)
-    private Integer quantity;
+    private Long quantity;
 
     @Column(nullable = false)
     private BigDecimal price;
